--- src/components/templates/twoColumnStat/index.ts
@@ -1,2 +1,0 @@
-export { default as TwoColumnStatTemplate } from './TwoColumnStat'
-export * from './TwoColumnStat.type'
