+++ src/components/templates/twoColumnStatText/TwoColumnStatText.type.ts
