+++ src/components/templates/twoColumnStatText/index.ts
@@ -1,0 +1,2 @@
+export { default as TwoColumnStatTemplate } from './TwoColumnStatText'
+export * from './TwoColumnStatText.type'
